disable loader on node 20
